refactor to remove InsertQueue
